Extract FilePreviewModal component from ChatPanel
Move the FilePreviewModal component into its own dedicated file to
improve code organization and reduce the size of ChatPanel.tsx
(from ~1650 lines to ~1530 lines).

This makes the modal component more reusable and improves maintainability
by isolating the file preview logic from the main chat panel.
